Fix tests for python 3
